BDSF Analysis Implementation

+++ src/analysis/bdsf_analysis.py
@@ -1,38 +1,36 @@
-from utils.data_utils import EvaluationDataset
-import utils.paths as paths
-from utils.paths import LOFAR_SUBSETS
-import argparse
-import multiprocessing as mp
-from concurrent.futures import ProcessPoolExecutor as PPEx
-from functools import partial
+import os
+import sys
 import signal
-import time
 import shutil
 import psutil
-import os
-
-import sys
-import os
 import pickle
+import logging
+import argparse
+import warnings
 import tempfile
-from collections.abc import Iterable
-from numbers import Number
+import traceback
+import functools
+import multiprocessing as mp
 from pathlib import Path
+from numbers import Number
+from collections.abc import Iterable
+from contextlib import contextmanager
+from concurrent.futures import ProcessPoolExecutor as PPEx
 
-from astropy.io import fits
 import bdsf
 import numpy as np
-from tqdm import tqdm
-from tqdm.contrib.concurrent import process_map
-import matplotlib.pyplot as plt
 import pandas as pd
+import matplotlib.pyplot as plt
+from tqdm import tqdm
+from astropy.io import fits
 
-import logging
-import functools
+import utils.paths as paths
+from utils.data_utils import EvaluationDataset
 
 
 def disable_logging(func):
     @functools.wraps(func)
+    #
     def wrapper(*args, **kwargs):
         logging.disable(logging.CRITICAL + 1)
         result = func(*args, **kwargs)
@@ -42,9 +40,9 @@ def disable_logging(func):
 
 
 # decorater used to block function printing to the console
-def block_printing(func):
+def redirect_printing(func):
     def func_wrapper(*args, **kwargs):
-        # block all printing to the console
+        # Redirect all printing to the console
         sys.stdout = open(os.devnull, 'w')
         # call the method in question
         try:
@@ -60,17 +58,21 @@ def block_printing(func):
     return func_wrapper
 
 
-@disable_logging
-@block_printing
-def bdsf_on_image(img: np.ndarray, ang_size=50, px_size=80):
+# @disable_logging
+# @block_printing
+def bdsf_on_image(img: np.ndarray, ang_size=120, px_size=80,
+                  id=None,
+                  tmpdir=paths.ANALYSIS_PARENT / 'tmp',
+                  logdir=None,
+                  ):
     '''
     Run bdsf on a single image.
     '''
     img = img.squeeze()
 
-    # Add small amount of noise, otherwise sigma-clipping algorithm called 
+    # Add small amount of noise, otherwise sigma-clipping algorithm called
     # within bdsf.process_image (functions.bstat) might not converge
-    z = np.random.normal(0, scale=min(img.max(), 1)*1e-5, size=img.shape)
+    z = np.random.normal(0, scale=min(img.max(), 1) * 1e-2, size=img.shape)
     img += z
 
     # Set up the header, which contains information required for bdsf
@@ -96,12 +98,11 @@ def bdsf_on_image(img: np.ndarray, ang_size=50, px_size=80):
     hdu = fits.PrimaryHDU(data=img, header=fits.Header(header_dict))
 
     # Create a temporary file
-    with tempfile.NamedTemporaryFile(suffix='.fits') as f:
+    with tempfile.NamedTemporaryFile(prefix=id, suffix='.fits', dir=tmpdir) as f:
 
         # Write the hdu to tmp fits file
         fits.HDUList([hdu]).writeto(f.name, overwrite=True)
 
-        # Run bdsf on the fits file
         img = bdsf.process_image(
             f.name,
             thresh_isl=1.5,
@@ -111,16 +112,23 @@ def bdsf_on_image(img: np.ndarray, ang_size=50, px_size=80):
             rms_map=False,
             thresh='hard',
             quiet=True,
+            debug=True,
         )
 
-        # Remove log file
-        os.remove(f'{f.name}.pybdsf.log')
+        # (Re)move log file
+        if logdir is not None:
+            os.rename(
+                tmpdir / f'{f.name}.pybdsf.log',
+                logdir / f'{id}.pybdsf.log'
+            )
+        else:
+            os.remove(f'{f.name}.pybdsf.log')
 
     # Return bdsf image object
     return img
 
 
-@block_printing
+@redirect_printing
 def catalogs_from_bdsf(img: bdsf.image.Image):
     cat_list = []
 
@@ -196,37 +204,54 @@ def writer(queue, q_pbar, fpaths, write_fns):
         if items == -1:
             print("Closing writer.")
             break
-        for item, fpath, write_fn in zip(items, fpaths, write_fns):
-            if item is None:
-                continue
-            write_fn(item, fpath)
 
-        # Write log
-        img_id = items[0]['Image_id']
-        with open(log_file, 'a' if i else 'w') as f:
-            f.write(f'{i},{img_id}\n')
-        i += 1
-        q_pbar.put(1)
+        elif items == 0:
+            i += 1
+            q_pbar.put(1)
+            continue
+
+        else:
+            for item, fpath, write_fn in zip(items, fpaths, write_fns):
+                if item is None:
+                    continue
+                write_fn(item, fpath)
+
+            # Write log
+            img_id = items[0]['Image_id']
+            with open(log_file, 'a' if i else 'w') as f:
+                f.write(f'{i},{img_id}\n')
+
+            i += 1
+            q_pbar.put(1)
+
 
 def kill_child_processes(parent_pid, sig=signal.SIGTERM):
+
     try:
         parent = psutil.Process(parent_pid)
     except psutil.NoSuchProcess:
         return
     children = parent.children(recursive=True)
+
     for process in children:
-        process.send_signal(sig)
+        try:
+            process.send_signal(sig)
+        except psutil.NoSuchProcess:
+            pass
 
 
-def bdsf_worker_func(img, image_id, q):
+def bdsf_worker_func(img, image_id, q, bdsf_kwargs):
+
+    # Handle keyboard interrupts
     def signal_handler(sig, frame):
         pid = os.getpid()
         print(f'Keyboard interrupt handeled by worker process {pid}.')
         return
 
     signal.signal(signal.SIGINT, signal_handler)
+
     # Run bdsf on the image
-    bdsf_img = bdsf_on_image(img)
+    bdsf_img = bdsf_on_image(img, id=image_id, **bdsf_kwargs)
 
     # bdsf dict:
     # Retrieve the desired attributes of the bdsf image object as dict
@@ -246,8 +271,16 @@ def bdsf_worker_func(img, image_id, q):
         if cat is not None:
             cat['Image_id'] = image_id
 
-    # Append to queue
+    # Append to writer queue
     q.put([bdsf_dict, *cats])
+
+
+def bdsf_wrapper(*args):
+    try:
+        bdsf_worker_func(*args)
+    except Exception as e:
+        args[2].put(0)
+        raise e
 
 
 def iterable_func_caller(func, args):
@@ -260,8 +293,9 @@ def progress_bar_worker(q_pbar, n_tot):
     def signal_handler(sig, frame):
         print('Keyboard interrupt ignored by progress bar worker.')
         q_pbar.put(-1)
+
     signal.signal(signal.SIGINT, signal_handler)
-    with tqdm(total=n_tot) as pbar:
+    with tqdm(total=n_tot, smoothing=0.1) as pbar:
         while True:
             sig = q_pbar.get()
             match sig:
@@ -283,30 +317,41 @@ def bdsf_run(
         names: Iterable[str] = None,
         override=False,
         max_workers=96,
+        **bdsf_kwargs,
 ):
     '''
     Run bdsf on a set of images and save the resulting list as pickle file.
     '''
+    # Set tmp directory
+    tmp_dir = paths.ANALYSIS_PARENT / 'tmp'
+    os.environ['TMPDIR'] = str(tmp_dir)
+    warnings.filterwarnings('ignore')
+
     # Set up the output folder
-    out_path = out_parent / out_folder
+    out_path = out_parent / out_folder / 'bdsf'
     out_path.mkdir(exist_ok=True)
 
     # Set up the output file paths
-    dicts_path = out_path / f'{out_folder}_bdsf_dicts'
+    dicts_path = out_path / f'dicts'
+    logs_path = out_path / f'logs'
+    err_path = out_path / f'errors'
     gaul_path = out_path / f'{out_folder}_bdsf_gaul.csv'
     srl_path = out_path / f'{out_folder}_bdsf_srl.csv'
 
     # If override is True, delete the files if they exist
     if override:
-        for fpath in [dicts_path, gaul_path, srl_path]:
+        for fpath in [
+            dicts_path, logs_path, err_path, gaul_path, srl_path,
+        ]:
             if fpath.exists():
                 fpath.unlink() if fpath.is_file() else shutil.rmtree(fpath)
 
     # Look for images already processed
     else:
         print('Looking for images already processed...')
-        # Open pickle file
+
         if dicts_path.exists():
+            # Open pickle files in dicts_path
             pkl_files = list(dicts_path.glob('*.pkl'))
             ids = np.array([f.stem for f in pkl_files])
             print(f'Removing {len(ids)} processed images from the list.')
@@ -316,13 +361,22 @@ def bdsf_run(
         else:
             print('No images processed yet.')
 
+    logs_path.mkdir(exist_ok=True)
+    err_path.mkdir(exist_ok=True)
     dicts_path.mkdir(exist_ok=True)
 
-    # Set up multiprocessing
+    # This will assign an id to every image/job
+    print('Getting image ids...')
+    image_ids = names if names is not None else range(len(imgs))
+    image_ids = [str(i) for i in image_ids]
+
+    # Prepare helper processes: writer and progress bar
     manager = mp.Manager()
     q = manager.Queue()
     q_pbar = manager.Queue()
     helper_pool = mp.Pool(2)
+
+    # Start the writer process
     w_res = helper_pool.apply_async(
         writer,
         (
@@ -332,18 +386,16 @@ def bdsf_run(
         )
     )
 
-    image_ids = names if names is not None else range(len(imgs))
-
-    # Init progress bar worker
+    # Start the progress bar process
     pbar_res = helper_pool.apply_async(
         progress_bar_worker,
         (q_pbar, len(imgs))
     )
-    # q_pbar.put(1)
 
-    # Run bdsf on the images
+    # Launch the worker pool
     with PPEx(max_workers=max_workers) as worker_pool:
 
+        # Handle keyboard interrupts
         def signal_handler(sig, frame):
             print('Keyboard interrupt. Closing pools.')
             worker_pool.shutdown(cancel_futures=True, wait=False)
@@ -352,36 +404,73 @@ def bdsf_run(
             helper_pool.close()
             helper_pool.join()
             kill_child_processes(os.getpid())
+            for f in tmp_dir.iterdir():
+                if f.suffix in ['.log', '.fits']:
+                    f.unlink()
             sys.exit(0)
 
         signal.signal(signal.SIGINT, signal_handler)
 
+        # Add log dir to bdsf kwargs
+        bdsf_kwargs['logdir'] = logs_path
+
+        # Run bdsf on the images
         print(f'Running bdsf on images. Parent process: {os.getpid()}.')
+        '''
         res = worker_pool.map(
             partial(iterable_func_caller, bdsf_worker_func),
-            zip(imgs, image_ids, [q] * len(imgs)),
+            zip(imgs, image_ids, [q] * len(imgs), [bdsf_kwargs] * len(imgs)),
             chunksize=10,
         )
-        [print(r) for r in res if r is not None]
+        '''
+        fut = []
+        for img, image_id in zip(imgs, image_ids):
+            f = worker_pool.submit(
+                bdsf_wrapper,
+                img, image_id, q, bdsf_kwargs
+            )
+            fut.append(f)
 
-    # Close the writer pool
+        # If any exceptions were raised, print to out file
+        err_count = 0
+        for i, f in enumerate(fut):
+            res = f.exception()
+            if res is not None:
+                with open(err_path / f'{image_ids[i]}.txt', 'w') as f:
+                    traceback.print_exception(res, file=f)
+                err_count += 1
+            else:
+                # Check if there is an error file from previous runs,
+                # if so remove.
+                err_file = err_path / f'{image_ids[i]}.txt'
+                if err_file.exists():
+                    err_file.unlink()
+                    
+        print(f'{err_count} errors were encountered - check error directory.')
+
+        # [print(r := f.result()) for f in fut if r is not None]
+        # [print(r.successful()) for r in res if r is not None]
+
+    # Finish helper processes
     print('Closing pools.')
     q.put(-1)
     q_pbar.put(-1)
     helper_pool.close()
     helper_pool.join()
+
+    # Check if writer was successful
     s = w_res.successful()
     print(f'Writer success: {s}')
     if not s:
         print(w_res.get())
-
-    print('Done.')
 
     # Check if progress bar worker was successful
     s = pbar_res.successful()
     print(f'Progress bar worker success: {s}')
     if not s:
         print(pbar_res.get())
+
+    print('Done.')
 
 
 def bdsf_plot(
@@ -400,6 +489,7 @@ def bdsf_plot(
 
 
 if __name__ == '__main__':
+
     parser = argparse.ArgumentParser()
     parser.add_argument(
         "-o", "--override", required=False,
@@ -408,23 +498,30 @@ if __name__ == '__main__':
     )
     arguments = parser.parse_args()
 
-    out_path = paths.ANALYSIS_PARENT /'Fmax_Context_Dropout'
+    '''
+    out_folder = paths.ANALYSIS_PARENT / 'Fmax_Context_Dropout'
     data_path = (
-        out_path /
-        'Fmax_Context_Dropout_samples_10000_guidance_strength=1.00e-01.pt'
+        out_folder
+        / 'Fmax_Context_Dropout_samples_10000_guidance_strength=1.00e-01.pt'
     )
+    '''
+    data_path = paths.LOFAR_SUBSETS['0-clip_unscaled']
 
     # Load the dataset
     dataset = EvaluationDataset(data_path)
+
+    # Scale Images
 
     # For testing: deterministic subset (use None for all images)
     n = None
 
     # Run bdsf on the images
     bdsf_out = bdsf_run(
-        dataset.data[:n],
-        out_folder=out_path,
+        # Call dataset for Transform
+        np.array([dataset[i] for i in range(n or len(dataset))]),
+        out_folder=data_path.stem,
         names=dataset.names[:n],
         override=arguments.override,
-        max_workers=80
+        max_workers=96,
+        ang_size=120,
     )
